Fix: refining PWM fails when match occurs at the end of the last sequence
refinePWMMotif -> scanPWMMatch -> bestPWMMatch -> DNAStringSet(seq)
  where seq <- subject(pasteSeq(seqs))
pasteSeq pads sequences with linker sequences of (default=10) + symbols
So that, when matches are extracted that start at negative positions, or
that span after the last position of a sequence, the sequences are
nicely embedded in linker symbols.
However, this fails, when positions extending beyond the end of the last
sequence are extracted.
This fixes that problem.
Likely, a similar problem exists for negative positions and the first
sequence.